add jquery autocomit plugin, viewlets
This hooks into the IEditAfterFieldset viewlet manager, so that the
required functionality is inserted more or less where it is required,
next to the cmfeditions viewlet. A method on the viewlet pulls the
possible change notes from the registry. json.dumps is used to simplify
this so we need not worry about weird characters in the configured
change notes.